requirements.txt
anthropic>=0.30.0
flask>=3.0.0
playwright>=1.44.0
gspread>=6.1.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
tenacity>=8.3.0
